renamed markov.py to gibbs.py
