+++ inference/mcmc/gibbs.py
